Refactoring the parsing of INI files
Instead of relying on regular expressions and quotations to handle multi-line
values, I am now using the "\" character at the end of a line to indicate the
value continues on to the next line. This is much easier to code and allows
values to be more flexible in what characters they contain. Stealing the
quotation character was a mistake.